JUnit zavrsen, izbrisani debug printovi

+++ Ticketplex/src/ticketplex/systemoperations/SOReservationRemove.java
@@ -7,14 +7,15 @@ import java.sql.SQLException;
 import db.Database;
 
 public class SOReservationRemove {
-	public static boolean execute(int reservation_id) {
+	public static boolean execute(int reservation_id, int user_id) {
 
 		Connection con = Database.getInstance().getConnection();
 
-		String sql = "DELETE FROM reservations WHERE id=?";
+		String sql = "DELETE FROM reservations WHERE id=? AND user_id = ?";
 
 		try (PreparedStatement pstmt = con.prepareStatement(sql)) {
 			pstmt.setInt(1, reservation_id);
+			pstmt.setInt(2, user_id);
 			return pstmt.executeUpdate() > 0 ? true : false;
 
 		} catch (SQLException e) {
